requirements.txt
beautifulsoup4==4.12.3
fastapi==0.110.1
fastembed==0.2.6
langchain==0.1.14
langchain-community==0.0.31
langchain-experimental==0.0.56
langchain-google-genai==1.0.1
python-decouple==3.8
qdrant-client==1.8.2
uvicorn==0.29.0